refactoring and layout adjusting
