+++ app/src/main/res/layout/activity_live_graph.xml
@@ -14,7 +14,7 @@
         android:layout_marginEnd="8dp"
         android:layout_marginStart="8dp"
         android:layout_marginTop="8dp"
-        android:text="Live Data Graph (MOCK DATA)"
+        android:text="@string/header_title"
         android:textColor="@android:color/black"
         android:textSize="16sp"
         app:layout_constraintEnd_toEndOf="parent"
